requirements.txt
streamlit==1.41.1
supabase==2.10.0
python-dotenv==1.0.1
pandas==2.2.3
